+++ examples/deployments.rs
@@ -1,16 +1,21 @@
 extern crate hyper;
 extern crate hubcaps;
+extern crate hyper_native_tls;
 
 use hyper::Client;
+use hyper::net::HttpsConnector;
+use hyper_native_tls::NativeTlsClient;
 use hubcaps::{Credentials, Github};
 use std::env;
 
 fn main() {
     match env::var("GITHUB_TOKEN").ok() {
         Some(token) => {
-            let github = Github::new(format!("hubcaps/{}", env!("CARGO_PKG_VERSION")),
-                                     Client::new(),
-                                     Credentials::Token(token));
+            let github =
+                Github::new(format!("hubcaps/{}", env!("CARGO_PKG_VERSION")),
+                            Client::with_connector(HttpsConnector::new(NativeTlsClient::new()
+                                .unwrap())),
+                            Credentials::Token(token));
             let repo = github.repo("softprops", "hubcaps");
             let deployments = repo.deployments();
             // let deploy = deployments.create(&DeploymentOptions::builder("master")
