+++ examples/orgs.rs
@@ -1,19 +1,24 @@
 extern crate env_logger;
 extern crate hyper;
 extern crate hubcaps;
+extern crate hyper_native_tls;
 
 use hyper::Client;
-use hubcaps::{Credentials, Github, OrganizationRepoListOptions};
-use hubcaps::repositories::OrgRepoType;
+use hyper::net::HttpsConnector;
+use hyper_native_tls::NativeTlsClient;
+use hubcaps::{Credentials, Github};
+use hubcaps::repositories::{OrgRepoType, OrganizationRepoListOptions};
 use std::env;
 
 fn main() {
     env_logger::init().unwrap();
     match env::var("GITHUB_TOKEN").ok() {
         Some(token) => {
-            let github = Github::new(format!("hubcaps/{}", env!("CARGO_PKG_VERSION")),
-                                     Client::new(),
-                                     Credentials::Token(token));
+            let github =
+                Github::new(format!("hubcaps/{}", env!("CARGO_PKG_VERSION")),
+                            Client::with_connector(HttpsConnector::new(NativeTlsClient::new()
+                                .unwrap())),
+                            Credentials::Token(token));
 
             let options =
                 OrganizationRepoListOptions::builder().repo_type(OrgRepoType::Forks).build();
