+++ rustfml.toml
@@ -1,0 +1,2 @@
+format_strings=false
+reorder_imports=true
